requirements.txt
async-generator==1.10
attrs==21.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chromedriver==2.24.1
click==8.0.3
cryptography==36.0.2
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
outcome==1.1.0
packaging==21.2
pdfkit==1.0.0
pluggy==1.0.0
py==1.10.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python-dotenv==0.21.0
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.26
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
Werkzeug==2.0.2
wsproto==1.1.0
WTForms==3.0.0
